Rollback rename and move of `Decompression`
Refactoring of `decompression` module will be done in a later PR
